(refactor codewhisperer): refactor 1, move supplemental context pojos to model file (#4029)
