+++ src/codewhisperer/util/editorContext.ts
@@ -11,13 +11,11 @@ import { getTabSizeSetting } from '../../shared/utilities/editorUtilities'
 import { TelemetryHelper } from './telemetryHelper'
 import { getLogger } from '../../shared/logger/logger'
 import { runtimeLanguageContext } from './runtimeLanguageContext'
-import {
-    CodeWhispererSupplementalContext,
-    fetchSupplementalContext,
-} from './supplementalContext/supplementalContextUtil'
+import { fetchSupplementalContext } from './supplementalContext/supplementalContextUtil'
 import { supplementalContextTimeoutInMs } from '../models/constants'
 import { getSelectedCustomization } from './customizationUtil'
 import { selectFrom } from '../../shared/utilities/tsUtils'
+import { CodeWhispererSupplementalContext } from '../models/model'
 
 let tabSize: number = getTabSizeSetting()
